+++ internal/bind/table_path_prefix.go
@@ -24,8 +24,8 @@ func (tablePathPrefix TablePathPrefix) NormalizePath(folderOrTable string) strin
 	}
 }
 
-func (tablePathPrefix TablePathPrefix) ToYdb(sql string, args ...interface{}) (
-	yql string, newArgs []interface{}, err error,
+func (tablePathPrefix TablePathPrefix) ToYdb(sql string, args ...any) (
+	yql string, newArgs []any, err error,
 ) {
 	buffer := xstring.Buffer()
 	defer buffer.Free()